train/requirements.txt
t5
tensorflow-datasets
tensorflow-text
tensorboard
cloud-tpu-profiler
tensorboard-plugin-profile
google-cloud-storage
termcolor